Regressions in 1.1.0pre (#247)
* fix(client): PILLOW raw image names

The RAW image formats using `int`s are named differently:

> Traceback (most recent call last):
>   File "vncdotool/rfb.py", line 1117, in dataReceived
>     self._handler()
>   File "vncdotool/rfb.py", line 1126, in _handleExpected
>     self._expected_handler(block, *self._expected_args, **self._expected_kwargs)
>   File "vncdotool/rfb.py", line 708, in _handleDecodeRAW
>     self.updateRectangle(x, y, width, height, block)
>   File "vncdotool/client.py", line 425, in updateRectangle
>     update = Image.frombytes('RGB', size, data, 'raw', self.image_mode)
>   File "PIL/Image.py", line 2882, in frombytes
>     im.frombytes(data, decoder_name, args)
>   File "PIL/Image.py", line 810, in frombytes
>     d = _getdecoder(self.mode, decoder_name, args)
>   File "PIL/Image.py", line 403, in _getdecoder
>     return decoder(mode, *args + extra)
> builtins.ValueError: unknown raw mode for given image mode

See source code of PILLOW for complete list:
<https://github.com/python-pillow/Pillow/blob/main/src/libImaging/Unpack.c>

Fixes: 15d1f70115ef4b30a3f089149d8812ee8eb2ce36

* fix(qemu): Fix Qemu extended key handling again

The PSEUDO_QEMU_EXTENDED_KEY_EVENT encoding may be part of a larger
set of update rectangles. Continue parsing the following rectangles
instead of going back to connection handling.

Actually call `_doConnection()` to handle either the next rectangle in
row or return to `_handleConnection` if it was the last rectangle.

Only call callback `commitUpdate()` if at least one real update
happened. Otherwise a PSEUDO_QEMU_EXTENDED_KEY_EVENT will trigger the
Deferred and `capture` will fail, as no real update has happened.

Fixes: 53627d814f97c5d14fd2cc9b879c4b967058b4cc

* fix(command): rms as float

+++ vncdotool/command.py
@@ -183,7 +183,7 @@ def build_command_list(
             factory.deferred.addCallback(client.captureScreen, filename, int(incremental_refreshes))
         elif cmd == 'expect':
             filename = args.pop(0)
-            rms = int(args.pop(0))
+            rms = float(args.pop(0))
             factory.deferred.addCallback(client.expectScreen, filename, rms)
         elif cmd == 'rcapture':
             filename = args.pop(0)
@@ -199,7 +199,7 @@ def build_command_list(
             filename = args.pop(0)
             x = int(args.pop(0))
             y = int(args.pop(0))
-            rms = int(args.pop(0))
+            rms = float(args.pop(0))
             factory.deferred.addCallback(client.expectRegion, filename, x, y, rms)
         elif cmd in ('pause', 'sleep'):
             duration = float(args.pop(0)) / warp
